Phase 2: Security and TLS Hardening (14 items)
Security fixes:
- Remove blanket #![allow(clippy::all)]; fix 544 clippy warnings
- TLS: default to https_only() upstream, add create_upstream_client()
- TLS: fallback to webpki roots when native certs unavailable
- TLS: wire skip_verify into NoVerifier (dangerous cert bypass)
- IPC key passed via temp file (0600) instead of env var
- IPC message validation expanded from 7 to 30+ variants
- CORS wildcard rejected in release builds
- Rate limiter check-and-reset now atomic via compare_exchange
- HSM PIN stored as Zeroizing<String> for secure cleanup
- Embedded Ed25519 key generates random key instead of zero-key
- 6 stub admin endpoints now return 501 NOT_IMPLEMENTED
- Config file chmod 0600 after --generatenewtoken
- Fix .cargo/config.toml target-dir affecting all builds
- Add rustls-native-certs dependency

Verification: cargo check, cargo check --features dns, 99/99 tests pass

+++ src/dns/compression.rs
@@ -153,7 +153,7 @@ impl DnsMessageDecompressor {
                     return Err("Invalid compression pointer".to_string());
                 }
 
-                let offset = ((len & 0x3F) as usize) << 8 | data[pos + 1] as usize;
+                let offset = (len & 0x3F) << 8 | data[pos + 1] as usize;
 
                 if !jumped {
                     jumped = true;
@@ -224,7 +224,7 @@ impl DnsMessageDecompressor {
                     return Err("Invalid pointer".to_string());
                 }
 
-                let offset = ((len & 0x3F) as usize) << 8 | data[pos + 1] as usize;
+                let offset = (len & 0x3F) << 8 | data[pos + 1] as usize;
 
                 pos = offset;
                 jumps += 1;
